+++ src/hooks/useLocalStorageItems.ts
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 
 import useAppSelector from './useAppSelector';
-import { useCustomCompareEffect } from 'use-custom-compare';
 
 import { HasKey } from '@/modules/common/CommonTypes';
 
@@ -25,59 +24,47 @@ function useLoadSave<T>(localStorageKey: string) {
     return { load, save };
 }
 
-const customCompare = (
-    prevDeps: React.DependencyList,
-    nextDeps: React.DependencyList
-): boolean => {
-    const prevItems: HasKey[] = prevDeps[0] || [];
-    const nextItems: HasKey[] = nextDeps[0] || [];
-
-    if (prevItems.length !== nextItems.length) {
-        return false;
-    }
-
-    return prevItems.reduce(
-        (areEqual: boolean, prevItem: HasKey, index: number): boolean => {
-            return areEqual && prevItem.key === nextItems[index].key;
-        },
-        true
-    );
-};
-
 function useLocalStorageItems<Type extends HasKey>(
     localStorageKey: string
 ): {
     items: Type[];
     addItem: (item: Type) => void;
+    updateItem: (item: Type) => void;
     removeItem: (item: Type) => void;
     clearItems: () => void;
 } {
     const { load, save } = useLoadSave<Type>(localStorageKey);
     const [items, setItems] = useState(load);
 
-    useCustomCompareEffect(
-        () => {
-            save(items);
-        },
-        [items],
-        customCompare
-    );
-
     const addItem = (item: Type) => {
-        setItems([...items, item]);
+        const newItems = [...items, item];
+        save(newItems);
+        setItems(newItems);
     };
 
     const removeItem = (item: Type) => {
-        setItems(items.filter((i) => i.key !== item.key));
+        const newItems = items.filter((i) => i.key !== item.key);
+        save(newItems);
+        setItems(newItems);
+    };
+
+    const updateItem = (item: Type) => {
+        const newItems = items.map((currentItem) =>
+            currentItem.key === item.key ? item : currentItem
+        );
+        save(newItems);
+        setItems(newItems);
     };
 
     const clearItems = () => {
+        save([]);
         setItems([]);
     };
 
     return {
         items,
         addItem,
+        updateItem,
         removeItem,
         clearItems,
     };