+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Show hyperparameters on the compare page (#434)
+-   Favorite and selected compute plans are updated when new data is available (#428)
+
 ## [0.15.0] - 2022-03-03
 
 ### Added